Delete Tags

+++ src/main/java/com/oktrueque/controller/TagController.java
@@ -1,6 +1,7 @@
 package com.oktrueque.controller;
 
 import com.oktrueque.model.Item;
+import com.oktrueque.model.ItemTag;
 import com.oktrueque.model.Tag;
 import com.oktrueque.service.ItemService;
 import com.oktrueque.service.ItemTagService;
@@ -9,9 +10,11 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 
 import javax.transaction.Transactional;
+import javax.xml.ws.Response;
 import java.util.List;
 
 @Controller
@@ -28,10 +31,17 @@ public class TagController {
         this.itemTagService = itemTagService;
     }
 
-    @RequestMapping(method = RequestMethod.GET, value = "/tags") //AutoComplete
+    @RequestMapping(method = RequestMethod.GET, value = "/tags/rest") //AutoComplete
     public ResponseEntity<List<Tag>> autocomplete(){
         List<Tag> tags = tagService.findAll();
         return new ResponseEntity<>(tags, HttpStatus.OK);
+    }
+
+    @RequestMapping(method = RequestMethod.GET, value = "/tags")
+    public String getTags(Model model){
+        List<Tag> tags = tagService.findAll();
+        model.addAttribute("tags", tags);
+        return "tags";
     }
 
     @RequestMapping(method = RequestMethod.POST, value = "/items/{id}/updateItemTags")
@@ -56,9 +66,14 @@ public class TagController {
     }
 
     @RequestMapping(method = RequestMethod.DELETE, value = "tags/{id}")
-    public String deleteTag(@PathVariable Long id){
-        tagService.deleteTag(id);
-        return "redirect:/tags";
+    public ResponseEntity<Boolean> deleteTag(@PathVariable Long id){
+        List<ItemTag> itemTags = itemTagService.getItemTagByTagId(id);
+        if(itemTags.size()>0){
+            return new ResponseEntity<>(false, HttpStatus.OK);
+        }else{
+            tagService.deleteTag(id);
+            return new ResponseEntity<>(true, HttpStatus.OK);
+        }
     }
 
 }
